Refactor fields to properties in Chart and Slide classes
Fields such as sdkChartPart, pGraphicFrame, and sdkSlidePart in Chart and Slide classes have been refactored to properties. This change makes these underlying instances of OpenXml elements more accessible and consistent with the overall codebase. In addition to this refactoring, getters have been added for these properties and their usage in the code has been updated accordingly.