Refactor code to simplify handling of shape fills
This commit includes two main updates. First, we've reduced some of the redundant code related to error messaging in the PresentationCore class. This simplifies the code and makes it easier to maintain. Secondly, the ShapeFillTests class has been refactored for efficiency. The process of setting the "NoFill" property on a shape has been streamlined to minimize the number of necessary operations.